Move replace operation to common util.
